--- .env.default
@@ -1,1 +1,0 @@
-JWT_SECRET=your-secret-key
