+++ src/main/java/com/back/global/websocket/config/WebSocketConfig.java
@@ -1,4 +1,4 @@
-package com.back.domain.websocket.config;
+package com.back.global.websocket.config;
 
 import org.springframework.context.annotation.Configuration;
 import org.springframework.messaging.simp.config.MessageBrokerRegistry;
